Move resolve_format function to formats module
